Move the schema functionality into a separate module.
